feat: fox farming provider -> programmatic UniV2 provider (#3863)
