+++ src/contracts/abis/IUniswapV2Router02.json
