+++ src/contracts/constants.ts
@@ -1,0 +1,21 @@
+export const WETH_TOKEN_CONTRACT_ADDRESS = '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
+export const FOX_TOKEN_CONTRACT_ADDRESS = '0xc770EEfAd204B5180dF6a14Ee197D99d808ee52d'
+
+// Staking contract addresses as string literals
+export const ETH_FOX_STAKING_CONTRACT_ADDRESS_V1 =
+  '0xdd80e21669a664bce83e3ad9a0d74f8dad5d9e72' as const
+export const ETH_FOX_STAKING_CONTRACT_ADDRESS_V2 =
+  '0xc54b9f82c1c54e9d4d274d633c7523f2299c42a0' as const
+export const ETH_FOX_STAKING_CONTRACT_ADDRESS_V3 =
+  '0x212ebf9fd3c10f371557b08e993eaab385c3932b' as const
+export const ETH_FOX_STAKING_CONTRACT_ADDRESS_V4 =
+  '0x24fd7fb95dc742e23dc3829d3e656feeb5f67fa0' as const
+export const ETH_FOX_STAKING_CONTRACT_ADDRESS_V5 =
+  '0xc14eaa8284feff79edc118e06cadbf3813a7e555' as const
+
+// Exported as a string literal for contract address discrimination purposes
+export const ETH_FOX_POOL_CONTRACT_ADDRESS = '0x470e8de2ebaef52014a47cb5e6af86884947f08c' as const
+
+// Exported as a string literal for contract address discrimination purposes
+export const UNISWAP_V2_ROUTER_02_CONTRACT_ADDRESS =
+  '0x7a250d5630b4cf539739df2c5dacb4c659f2488d' as const
