+++ src/features/defi/contexts/DefiManagerProvider/DefiCommon.ts
@@ -11,7 +11,8 @@ export enum DefiProvider {
   Idle = 'idle',
   Yearn = 'yearn',
   ShapeShift = 'ShapeShift',
-  FoxFarming = 'ShapeShift Farming',
+  EthFoxStaking = 'ETH/FOX Staking',
+  UniV2 = 'Uniswap V2',
   Cosmos = 'Cosmos',
   Osmosis = 'Osmosis',
   ThorchainSavers = 'THORChain Savers',
