--- src/plugins/foxPage/const.ts
@@ -1,3 +1,0 @@
-export const FOX_TOKEN_CONTRACT_ADDRESS = '0xc770EEfAd204B5180dF6a14Ee197D99d808ee52d'
-export const WETH_TOKEN_CONTRACT_ADDRESS = '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
-export const TRADING_FEE_RATE = 0.003
