+++ src/state/slices/opportunitiesSlice/opportunitiesSlice.test.ts
@@ -40,7 +40,7 @@ describe('opportunitiesSlice', () => {
               id: foxEthLpAssetId,
               name: 'ETH/FOX Pool',
               underlyingAssetId: foxEthLpAssetId,
-              provider: DefiProvider.FoxFarming,
+              provider: DefiProvider.UniV2,
               tvl: '424242',
               apy: '0.42',
               type: DefiType.LiquidityPool,
@@ -67,7 +67,7 @@ describe('opportunitiesSlice', () => {
               id: foxEthLpAssetId,
               name: 'ETH/FOX Pool',
               underlyingAssetId: foxEthLpAssetId,
-              provider: DefiProvider.FoxFarming,
+              provider: DefiProvider.UniV2,
               tvl: '424242',
               apy: '0.42',
               type: DefiType.LiquidityPool,
@@ -92,7 +92,7 @@ describe('opportunitiesSlice', () => {
               id: foxEthLpAssetId,
               name: 'ETH/FOX Pool',
               underlyingAssetId: foxEthLpAssetId,
-              provider: DefiProvider.FoxFarming,
+              provider: DefiProvider.UniV2,
               tvl: '424242',
               apy: '0.42',
               type: DefiType.LiquidityPool,
