--- src/state/slices/opportunitiesSlice/resolvers/foxFarming/index.ts
@@ -1,320 +1,0 @@
-import { ethAssetId, ethChainId, foxAssetId, fromAccountId, fromAssetId } from '@shapeshiftoss/caip'
-import type { MarketData } from '@shapeshiftoss/types'
-import { HistoryTimeframe } from '@shapeshiftoss/types'
-import { Fetcher, Token } from '@uniswap/sdk'
-import dayjs from 'dayjs'
-import { DefiProvider, DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
-import { FOX_TOKEN_CONTRACT_ADDRESS, WETH_TOKEN_CONTRACT_ADDRESS } from 'plugins/foxPage/const'
-import {
-  calculateAPRFromToken0,
-  getEthersProvider,
-  makeTotalLpApr,
-  rewardRatePerToken,
-} from 'plugins/foxPage/utils'
-import { bn, bnOrZero } from 'lib/bignumber/bignumber'
-import { toBaseUnit } from 'lib/math'
-import type { ReduxState } from 'state/reducer'
-import type { AssetsState } from 'state/slices/assetsSlice/assetsSlice'
-import { marketData } from 'state/slices/marketDataSlice/marketDataSlice'
-import type { PortfolioAccountBalancesById } from 'state/slices/portfolioSlice/portfolioSliceCommon'
-import { selectPortfolioLoadingStatusGranular } from 'state/slices/portfolioSlice/selectors'
-import { selectMarketDataById, selectPortfolioAccountBalances } from 'state/slices/selectors'
-
-import {
-  assertIsFoxEthStakingContractAddress,
-  foxEthLpAssetId,
-  foxEthLpAssetIds,
-  foxEthLpContractAddress,
-  foxEthPair,
-  foxEthStakingIds,
-  STAKING_ID_TO_VERSION,
-} from '../../constants'
-import type {
-  GetOpportunityIdsOutput,
-  GetOpportunityMetadataOutput,
-  GetOpportunityUserStakingDataOutput,
-} from '../../types'
-import { serializeUserStakingId } from '../../utils'
-import type { OpportunityMetadataResolverInput, OpportunityUserDataResolverInput } from '../types'
-import { fetchPairData, getOrCreateContract } from './contractManager'
-
-export const foxFarmingLpMetadataResolver = async ({
-  opportunityId,
-  opportunityType,
-  reduxApi,
-}: OpportunityMetadataResolverInput): Promise<{
-  data: GetOpportunityMetadataOutput
-}> => {
-  const { dispatch, getState } = reduxApi
-  const { assetReference: contractAddress } = fromAssetId(opportunityId)
-  const state: any = getState() // ReduxState causes circular dependency
-  const assets: AssetsState = state.assets
-  const ethMarketData: MarketData = selectMarketDataById(state, ethAssetId)
-
-  if (bnOrZero(ethMarketData?.price).isZero()) {
-    throw new Error(`Market data not ready for ${ethAssetId}`)
-  }
-
-  const ethPrecision = assets.byId[ethAssetId]?.precision
-  const lpAssetPrecision = assets.byId[foxEthLpAssetId]?.precision
-  const ethPrice = ethMarketData.price
-  const ethersProvider = getEthersProvider()
-  const uniV2LPContract = getOrCreateContract(contractAddress as typeof foxEthLpContractAddress)
-  const pair = await fetchPairData(
-    new Token(0, WETH_TOKEN_CONTRACT_ADDRESS, 18),
-    new Token(0, FOX_TOKEN_CONTRACT_ADDRESS, 18),
-    Fetcher.fetchPairData,
-    ethersProvider,
-  )
-
-  const blockNumber = await ethersProvider.getBlockNumber()
-
-  const calculatedApy = await calculateAPRFromToken0({
-    token0Decimals: pair.token0.decimals,
-    token0Reserves: pair.reserve0,
-    blockNumber,
-    uniswapLPContract: uniV2LPContract,
-  })
-  const apy = bnOrZero(calculatedApy).div(100).toString()
-  const reserves = await uniV2LPContract.getReserves()
-
-  // Getting the ratio of the LP token for each asset
-  const totalSupply = (await uniV2LPContract.totalSupply()).toString()
-  const foxReserves = bnOrZero(bnOrZero(reserves[1].toString()).toString())
-  const ethReserves = bnOrZero(bnOrZero(reserves[0].toString()).toString())
-  const ethPoolRatio = ethReserves.div(totalSupply).toString()
-  const foxPoolRatio = foxReserves.div(totalSupply).toString()
-  // Amount of Eth in liquidity pool
-  const ethInReserve = bnOrZero(reserves?.[0]?.toString()).div(`1e${ethPrecision}`)
-
-  // TODO(gomes): This is a hackish way we were previously doing this, getting the ETH liquidity * 2 will get us close to the real fiat TVL but we can do better
-  // Total market cap of liquidity pool in usdc.
-  // Multiplied by 2 to show equal amount of eth and fox.
-  const totalLiquidityFiat = ethInReserve.times(ethPrice).times(2)
-  const tvl = totalLiquidityFiat.toString()
-  const price = bnOrZero(tvl)
-    .div(bnOrZero(totalSupply.toString()).div(`1e${lpAssetPrecision}`))
-    .toString()
-
-  const lpMarketData = {
-    [foxEthLpAssetId]: { price, marketCap: '0', volume: '0', changePercent24Hr: 0 },
-  }
-  // hacks for adding lp price and price history
-  dispatch(marketData.actions.setCryptoMarketData(lpMarketData))
-  Object.values(HistoryTimeframe).forEach(timeframe => {
-    dispatch(
-      marketData.actions.setCryptoPriceHistory({
-        data: [{ price: bnOrZero(price).toNumber(), date: 0 }],
-        args: { timeframe, assetId: foxEthLpAssetId },
-      }),
-    )
-  })
-
-  const data = {
-    byId: {
-      [opportunityId]: {
-        apy,
-        assetId: opportunityId,
-        id: opportunityId,
-        provider: DefiProvider.FoxFarming,
-        tvl,
-        type: DefiType.LiquidityPool,
-        underlyingAssetId: foxEthLpAssetId,
-        underlyingAssetIds: foxEthPair,
-        underlyingAssetRatiosBaseUnit: [
-          toBaseUnit(ethPoolRatio.toString(), assets.byId[foxEthPair[0]]?.precision ?? 0),
-          toBaseUnit(foxPoolRatio.toString(), assets.byId[foxEthPair[1]]?.precision ?? 0),
-        ] as const,
-        name: 'ETH/FOX Pool',
-      },
-    },
-
-    type: opportunityType,
-  } as const
-
-  return { data }
-}
-
-export const foxFarmingStakingMetadataResolver = async ({
-  opportunityId,
-  opportunityType,
-  reduxApi,
-}: OpportunityMetadataResolverInput): Promise<{
-  data: GetOpportunityMetadataOutput
-}> => {
-  const { getState } = reduxApi
-  const state: any = getState() // ReduxState causes circular dependency
-  const assets: AssetsState = state.assets
-  const lpAssetPrecision = assets.byId[foxEthLpAssetId]?.precision ?? 0
-  const foxPrecision = assets.byId[foxAssetId]?.precision ?? 18
-  const ethPrecision = assets.byId[ethAssetId]?.precision ?? 18
-  const lpTokenMarketData: MarketData = selectMarketDataById(state, foxEthLpAssetId)
-  const lpTokenPrice = lpTokenMarketData?.price
-
-  const { assetReference: contractAddress } = fromAssetId(opportunityId)
-
-  if (bnOrZero(lpTokenPrice).isZero()) {
-    throw new Error(`Market data not ready for ${foxEthLpAssetId}`)
-  }
-
-  const ethersProvider = getEthersProvider()
-  assertIsFoxEthStakingContractAddress(contractAddress)
-  const foxFarmingContract = getOrCreateContract(contractAddress)
-  const uniV2LPContract = getOrCreateContract(foxEthLpContractAddress)
-
-  // tvl
-  const totalSupply = await foxFarmingContract.totalSupply()
-  const tvl = bnOrZero(totalSupply.toString())
-    .div(bn(10).pow(lpAssetPrecision))
-    .times(lpTokenPrice)
-    .toFixed(2)
-
-  // apr
-  const foxRewardRatePerTokenV5 = await rewardRatePerToken(foxFarmingContract)
-  const pair = await Fetcher.fetchPairData(
-    new Token(0, WETH_TOKEN_CONTRACT_ADDRESS, ethPrecision),
-    new Token(0, FOX_TOKEN_CONTRACT_ADDRESS, foxPrecision),
-    ethersProvider,
-  )
-
-  // Getting the ratio of the LP token for each asset
-  const reserves = await uniV2LPContract.getReserves()
-  const lpTotalSupply = (await uniV2LPContract.totalSupply()).toString()
-  const foxReserves = bnOrZero(bnOrZero(reserves[1].toString()).toString())
-  const ethReserves = bnOrZero(bnOrZero(reserves[0].toString()).toString())
-  const ethPoolRatio = ethReserves.div(lpTotalSupply).toString()
-  const foxPoolRatio = foxReserves.div(lpTotalSupply).toString()
-
-  const totalSupplyV2 = await uniV2LPContract.totalSupply()
-
-  const token1PoolReservesEquivalent = bnOrZero(pair.reserve1.toFixed())
-    .times(2) // Double to get equivalent of both sides of pool
-    .times(bn(10).pow(pair.token1.decimals)) // convert to base unit value
-
-  const foxEquivalentPerLPToken = token1PoolReservesEquivalent
-    .div(bnOrZero(totalSupplyV2.toString()))
-    .times(bn(10).pow(pair.token1.decimals)) // convert to base unit value
-    .toString()
-  const apy = bnOrZero(makeTotalLpApr(foxRewardRatePerTokenV5, foxEquivalentPerLPToken))
-    .div(100)
-    .toString()
-
-  const timeStamp = await foxFarmingContract.periodFinish()
-  const expired =
-    timeStamp.toNumber() === 0 ? false : dayjs().isAfter(dayjs.unix(timeStamp.toNumber()))
-  const version = STAKING_ID_TO_VERSION[opportunityId]
-
-  const data = {
-    byId: {
-      [opportunityId]: {
-        apy,
-        assetId: opportunityId,
-        id: opportunityId,
-        provider: DefiProvider.FoxFarming,
-        tvl,
-        type: DefiType.Staking,
-        underlyingAssetId: foxEthLpAssetId,
-        underlyingAssetIds: foxEthPair,
-        underlyingAssetRatiosBaseUnit: [
-          toBaseUnit(ethPoolRatio.toString(), assets.byId[foxEthPair[0]]?.precision ?? 0),
-          toBaseUnit(foxPoolRatio.toString(), assets.byId[foxEthPair[1]]?.precision ?? 0),
-        ] as const,
-        expired,
-        name: 'Fox Farming',
-        version,
-      },
-    },
-    type: opportunityType,
-  } as const
-
-  return { data }
-}
-
-export const foxFarmingLpUserDataResolver = ({
-  opportunityId,
-  opportunityType: _opportunityType,
-  accountId,
-  reduxApi,
-}: OpportunityUserDataResolverInput): Promise<void> => {
-  const { chainId: accountChainId } = fromAccountId(accountId)
-  // Looks the same as the happy path but isn't, we won't hit this as a guard with non-Ethereum account ChainIds
-  if (accountChainId !== ethChainId) return Promise.resolve()
-
-  const { getState } = reduxApi
-  const state: ReduxState = getState() as any
-  const portfolioLoadingStatusGranular = selectPortfolioLoadingStatusGranular(state)
-
-  // Reject RTK query if account portfolio data is granularily loading
-  if (portfolioLoadingStatusGranular?.[accountId] === 'loading')
-    throw new Error(`Portfolio data not loaded for ${accountId}`)
-
-  const balances: PortfolioAccountBalancesById = selectPortfolioAccountBalances(state)
-
-  const hasPortfolioData = Boolean(balances[accountId][opportunityId])
-
-  // Reject RTK query if there's no account portfolio data for this LP token
-  if (!hasPortfolioData) {
-    throw new Error('no portfolio data')
-  }
-
-  // All checks passed, resolve the promise so we continue the RTK query execution and populate LP/Account IDs
-  return Promise.resolve()
-}
-
-export const foxFarmingStakingUserDataResolver = async ({
-  opportunityId,
-  opportunityType,
-  accountId,
-  reduxApi,
-}: OpportunityUserDataResolverInput): Promise<{ data: GetOpportunityUserStakingDataOutput }> => {
-  const { chainId: accountChainId } = fromAccountId(accountId)
-  if (accountChainId !== ethChainId)
-    return {
-      data: {
-        byId: {},
-      },
-    }
-  const { getState } = reduxApi
-  const state: any = getState() // ReduxState causes circular dependency
-  const lpTokenMarketData: MarketData = selectMarketDataById(state, foxEthLpAssetId)
-  const lpTokenPrice = lpTokenMarketData?.price
-
-  const { assetReference: contractAddress } = fromAssetId(opportunityId)
-  const { account: accountAddress } = fromAccountId(accountId)
-
-  if (bnOrZero(lpTokenPrice).isZero()) {
-    throw new Error(`Market data not ready for ${foxEthLpAssetId}`)
-  }
-
-  assertIsFoxEthStakingContractAddress(contractAddress)
-
-  const foxFarmingContract = getOrCreateContract(contractAddress)
-
-  const stakedBalance = await foxFarmingContract.balanceOf(accountAddress)
-  const earned = await foxFarmingContract.earned(accountAddress)
-  const stakedAmountCryptoBaseUnit = bnOrZero(stakedBalance.toString()).toString()
-  const rewardsAmountsCryptoBaseUnit = [earned.toString()] as [string]
-
-  const userStakingId = serializeUserStakingId(accountId, opportunityId)
-
-  const data = {
-    byId: {
-      [userStakingId]: {
-        userStakingId,
-        stakedAmountCryptoBaseUnit,
-        rewardsAmountsCryptoBaseUnit,
-      },
-    },
-    type: opportunityType,
-  }
-
-  return { data }
-}
-
-export const foxFarmingLpOpportunityIdsResolver = (): Promise<{
-  data: GetOpportunityIdsOutput
-}> => Promise.resolve({ data: [...foxEthLpAssetIds] })
-
-export const foxFarmingStakingOpportunityIdsResolver = (): Promise<{
-  data: GetOpportunityIdsOutput
-}> => Promise.resolve({ data: [...foxEthStakingIds] })
