+++ src/state/slices/opportunitiesSlice/resolvers/index.ts
@@ -6,13 +6,10 @@ import {
   cosmosSdkStakingOpportunitiesUserDataResolver,
 } from './cosmosSdk'
 import {
-  foxFarmingLpMetadataResolver,
-  foxFarmingLpOpportunityIdsResolver,
-  foxFarmingLpUserDataResolver,
-  foxFarmingStakingMetadataResolver,
-  foxFarmingStakingOpportunityIdsResolver,
-  foxFarmingStakingUserDataResolver,
-} from './foxFarming'
+  ethFoxStakingMetadataResolver,
+  ethFoxStakingOpportunityIdsResolver,
+  ethFoxStakingUserDataResolver,
+} from './ethFoxStaking'
 import {
   foxyStakingOpportunitiesMetadataResolver,
   foxyStakingOpportunitiesUserDataResolver,
@@ -34,15 +31,22 @@ import {
   thorchainSaversStakingOpportunitiesUserDataResolver,
 } from './thorchainsavers'
 import {
+  uniV2LpLpOpportunityIdsResolver,
+  uniV2LpMetadataResolver,
+  uniV2LpUserDataResolver,
+} from './uniV2'
+import {
   yearnStakingOpportunitiesMetadataResolver,
   yearnStakingOpportunitiesUserDataResolver,
   yearnStakingOpportunityIdsResolver,
 } from './yearn'
 
 export const DefiProviderToMetadataResolverByDeFiType = {
-  [`${DefiProvider.FoxFarming}`]: {
-    [`${DefiType.LiquidityPool}`]: foxFarmingLpMetadataResolver,
-    [`${DefiType.Staking}`]: foxFarmingStakingMetadataResolver,
+  [`${DefiProvider.UniV2}`]: {
+    [`${DefiType.LiquidityPool}`]: uniV2LpMetadataResolver,
+  },
+  [`${DefiProvider.EthFoxStaking}`]: {
+    [`${DefiType.Staking}`]: ethFoxStakingMetadataResolver,
   },
 }
 
@@ -86,9 +90,11 @@ export const DefiProviderToOpportunitiesUserDataResolverByDeFiType = {
 }
 
 export const DefiProviderToOpportunityIdsResolverByDeFiType = {
-  [`${DefiProvider.FoxFarming}`]: {
-    [`${DefiType.LiquidityPool}`]: foxFarmingLpOpportunityIdsResolver,
-    [`${DefiType.Staking}`]: foxFarmingStakingOpportunityIdsResolver,
+  [`${DefiProvider.UniV2}`]: {
+    [`${DefiType.LiquidityPool}`]: uniV2LpLpOpportunityIdsResolver,
+  },
+  [`${DefiProvider.EthFoxStaking}`]: {
+    [`${DefiType.Staking}`]: ethFoxStakingOpportunityIdsResolver,
   },
   [`${DefiProvider.Idle}`]: {
     [`${DefiType.Staking}`]: idleStakingOpportunityIdsResolver,
@@ -111,9 +117,11 @@ export const DefiProviderToOpportunityIdsResolverByDeFiType = {
 }
 
 export const DefiProviderToUserDataResolverByDeFiType = {
-  [`${DefiProvider.FoxFarming}`]: {
-    [`${DefiType.LiquidityPool}`]: foxFarmingLpUserDataResolver,
-    [`${DefiType.Staking}`]: foxFarmingStakingUserDataResolver,
+  [`${DefiProvider.UniV2}`]: {
+    [`${DefiType.LiquidityPool}`]: uniV2LpUserDataResolver,
+  },
+  [`${DefiProvider.EthFoxStaking}`]: {
+    [`${DefiType.Staking}`]: ethFoxStakingUserDataResolver,
   },
   [`${DefiProvider.Osmosis}`]: {
     [`${DefiType.LiquidityPool}`]: osmosisLpUserDataResolver,
