+++ src/state/slices/opportunitiesSlice/resolvers/uniV2/constants.ts
@@ -1,0 +1,1 @@
+export const TRADING_FEE_RATE = 0.003
