+++ src/state/slices/opportunitiesSlice/selectors/selectors.test.ts
@@ -158,7 +158,7 @@ describe('opportunitiesSlice selectors', () => {
       assetId: foxEthLpAssetId,
       id: foxEthLpAssetId,
       name: 'FOX Farming',
-      provider: DefiProvider.FoxFarming,
+      provider: DefiProvider.UniV2,
       tvl: '424242',
       apy: '0.42',
       type: DefiType.LiquidityPool,
@@ -222,7 +222,7 @@ describe('opportunitiesSlice selectors', () => {
         id: 'eip155:1/erc20:0x470e8de2ebaef52014a47cb5e6af86884947f08c',
         userStakingId: 'eip155:1:0xgomes*eip155:1:0xStakingContractTwo',
         name: 'FOX Farming',
-        provider: DefiProvider.FoxFarming,
+        provider: DefiProvider.UniV2,
         rewardsAmountsCryptoBaseUnit: ['420000000000000000000'],
         stakedAmountCryptoBaseUnit: '1337',
         tvl: '424242',
@@ -244,7 +244,7 @@ describe('opportunitiesSlice selectors', () => {
         id: 'eip155:1/erc20:0x470e8de2ebaef52014a47cb5e6af86884947f08c',
         userStakingId: 'eip155:1:0xgomes*eip155:1:0xStakingContractOne',
         name: 'FOX Farming',
-        provider: DefiProvider.FoxFarming,
+        provider: DefiProvider.UniV2,
         stakedAmountCryptoBaseUnit: '4',
         rewardsAmountsCryptoBaseUnit: ['3000000000000000000'] as [string],
         tvl: '424242',
@@ -267,7 +267,7 @@ describe('opportunitiesSlice selectors', () => {
           assetId: mockStakingContractTwo,
           name: 'FOX Farming',
           id: mockStakingContractTwo,
-          provider: DefiProvider.FoxFarming,
+          provider: DefiProvider.UniV2,
           tvl: '91283233211',
           type: DefiType.LiquidityPool,
           underlyingAssetIds: foxEthPair,
@@ -325,7 +325,7 @@ describe('opportunitiesSlice selectors', () => {
             assetId: mockStakingContractTwo,
             id: mockStakingContractTwo,
             name: 'FOX Farming',
-            provider: DefiProvider.FoxFarming,
+            provider: DefiProvider.UniV2,
             rewardsAmountsCryptoBaseUnit: ['420000000000000000000'] as [string],
             stakedAmountCryptoBaseUnit: '1337',
             tvl: '91283233211',
@@ -343,7 +343,7 @@ describe('opportunitiesSlice selectors', () => {
             assetId: mockStakingContractTwo,
             id: mockStakingContractTwo,
             name: 'FOX Farming',
-            provider: DefiProvider.FoxFarming,
+            provider: DefiProvider.UniV2,
             rewardsAmountsCryptoBaseUnit: ['1000000000000000000'] as [string],
             stakedAmountCryptoBaseUnit: '100',
             tvl: '91283233211',
@@ -370,7 +370,7 @@ describe('opportunitiesSlice selectors', () => {
           id: mockStakingContractTwo,
           userStakingId: 'eip155:1:0xcatpuccino*eip155:1:0xStakingContractTwo',
           name: 'FOX Farming',
-          provider: DefiProvider.FoxFarming,
+          provider: DefiProvider.UniV2,
           tvl: '91283233211',
           type: DefiType.LiquidityPool,
           underlyingAssetId: foxEthLpAssetId,
